gui: use relm factories for questions tabs
It's better, but is has some bugs for now. The position of tabs are not
synced with the factory vector.

This issue should be solved in Relm4's side, or we should find a hack to
solve it. If it wasn't solved this commit should be reverted before a beta
release.

Read More:
    https://github.com/Relm4/Relm4/issues/573